properly use alternative names

+++ paciofs-server/src/test/scripts/gen-certs.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
+if [ ! -e "$(pwd)/gen-certs.sh" ]; then
+  echo "Must be run from this directory"
+  exit 1
+fi
+
 # see openssl.cnf
 mkdir -p ./paciofsCA
 
 # clean
+rm *.p12
+rm *.pass
 rm *.pem
 rm -rf ./paciofsCA/*
 
@@ -13,23 +20,43 @@ touch ./paciofsCA/index.txt
 # serial number file
 echo "01" > ./paciofsCA/serial
 
-# generate private keys and certificate signing requests
+# create self-signed certificate authority
 mkdir ./paciofsCA/private
-openssl req -config ./openssl.cnf -new -out ca.csr -newkey rsa:2048 -keyout ./paciofsCA/private/cakey.pem -subj "/C=DE/ST=Berlin/L=Berlin/O=ZIB/OU=DAS/CN=ca.paciofs.zib.de" -nodes -sha256 -batch
-openssl req -config ./openssl.cnf -new -out server.csr -newkey rsa:2048 -keyout serverkey.pem -subj "/C=DE/ST=Berlin/L=Berlin/O=ZIB/OU=DAS/CN=server.paciofs.zib.de" -nodes -sha256 -batch
-openssl req -config ./openssl.cnf -new -out client.csr -newkey rsa:2048 -keyout clientkey.pem -subj "/C=DE/ST=Berlin/L=Berlin/O=ZIB/OU=DAS/CN=client.paciofs.zib.de" -nodes -sha256 -batch
+openssl genrsa -out ./paciofsCA/private/cakey.pem 2048
+openssl req -new -x509 -key ./paciofsCA/private/cakey.pem -out ./paciofsCA/cacert.pem -subj "/C=DE/ST=Berlin/L=Berlin/O=ZIB/OU=DAS/CN=ca.paciofs.zib.de" -days 365 -batch
 
-# create self-signed root certificate
-openssl x509 -req -in ca.csr -signkey ./paciofsCA/private/cakey.pem -out ./paciofsCA/cacert.pem -sha256
-rm ca.csr
+# generate private keys and certificate signing request
+openssl req -config ./openssl.cnf -new -out client.csr -newkey rsa:2048 -keyout clientkey.pem -subj "/C=DE/ST=Berlin/L=Berlin/O=ZIB/OU=DAS/CN=client.paciofs.zib.de" -nodes -sha256 -batch
+openssl req -config ./openssl.cnf -new -out server.csr -newkey rsa:2048 -keyout serverkey.pem -subj "/C=DE/ST=Berlin/L=Berlin/O=ZIB/OU=DAS/CN=server.paciofs.zib.de" -nodes -sha256 -batch
 
 # sign server and client certificates
 mkdir ./paciofsCA/newcerts
-openssl ca -config ./openssl.cnf -in server.csr -out server.pem -md sha256 -batch
-rm server.csr
-openssl ca -config ./openssl.cnf -in client.csr -out client.pem -md sha256 -batch
+
+# append alternative name before signing
+echo "DNS.1 = client.paciofs.zib.de" >> ./openssl.cnf
+openssl ca -config ./openssl.cnf -in client.csr -out client.pem -md sha256 -batch -days 365
 rm client.csr
 
-# create the chain for the client
-cat server.pem > chain.pem
-cat ./paciofsCA/cacert.pem >> chain.pem
+# remove again
+tail -n 1 ./openssl.cnf | wc -c | xargs -I {} truncate ./openssl.cnf -s -{}
+
+# same for server
+echo "DNS.1 = server.paciofs.zib.de" >> ./openssl.cnf
+echo "DNS.2 = localhost" >> ./openssl.cnf
+echo "DNS.3 = 127.0.0.1" >> ./openssl.cnf
+echo "DNS.4 = $(uname -n)" >> ./openssl.cnf
+openssl ca -config ./openssl.cnf -in server.csr -out server.pem -md sha256 -batch -days 365
+rm server.csr
+tail -n 4 ./openssl.cnf | wc -c | xargs -I {} truncate ./openssl.cnf -s -{}
+
+# export server certificate to p12
+pwgen -Bs 10 1 > server.p12.pass
+openssl pkcs12 -export -in server.pem -inkey serverkey.pem -out server.p12 -name "server" -passout file:server.p12.pass
+
+# install certificates
+cp server.p12 ../resources/certs/
+cp server.p12.pass ../resources/certs/
+
+cp ./paciofsCA/cacert.pem ../../../../paciofs-fuse/test/certs/
+cp client.pem ../../../../paciofs-fuse/test/certs/
+cp clientkey.pem ../../../../paciofs-fuse/test/certs/
